Avoid warnings with GCC 4.1.2

+++ gatb-core/src/gatb/tools/collections/impl/Hash16.hpp
@@ -80,7 +80,7 @@ public:
 
         tai         = (1LL << tai_Hash16);
         mask        = tai-1 ;
-        max_nb_elem = 0.8*sizeMB*1024LL*1024LL /sizeof(cell);
+        max_nb_elem = (u_int64_t) (0.8*sizeMB*1024LL*1024LL /sizeof(cell));
         datah       = (cell_ptr_t *) _memory.malloc( tai * sizeof(cell_ptr_t));  //create hashtable
 
         _memory.memset (datah,0, tai * sizeof(cell_ptr_t));
